fix: Correct bitbucket Author, Org, and Repo.
The issue in jenkins-x/jx#664 is caused by some mismatching of the PR's
Author and the "username" in the various Bitbucket APIs. In the case
where the repository is part of an organization the "username" is
actually the organization name. Likewise, if the PR author does not
match the repository username the same issue would occur.

This commit updates the `bitbucket_cloud.go` provider to use the
organization name as the PR `Owner` and the short repository name as the
`Repo`. It also adds the `Author` to the returned PR and uses the
proviers `UserInfo` method to obtain the full author information.