fix(CI) test running CI on new cluster
after switching to the new cluster for CI we had an issue looking up the git details as we don't mount the secret anymore.  The problem was the existing code found a local `gitAuth.yaml` file but also a pipeline secrets in the test cluster.  This change avoids them being merged.