fix:ijx location was wrong when run on windows with cygwin.
when cygwin `where` was on the path in windows the path to jx (if it was
in the path) would be a cygwin path which was doomed to failure.
bogus on windows

```
 Downloading
 https://github.com/jenkins-x/jx/releases/download/v1.3.492/jx-windows-amd64.tar.gz
 to \cygdrive\c\tools\jx.tgz...
```

if cygwin was not installed the path was also bogus as the `where`
command does not exist and as such it would try and install it into
%HOME%\.jx\bin\jx which would mean the command would not be updated
(unless that was where it was installed.

So this actually assumes that we are running the jx command and we will
upgrade the current binary (which seems like a fair assumption!)