feat(serve-buildnumbers): Add new cmd to serve build numbers over HTTP interface.
Add a new command `jx serve-buildnumbers` that will run an HTTP interface to serve up build numbers
similar to Prow's tot. The interface is the same as tot's (i.e. /vend/<id>) to get the next sequential
build number.
As used in Jenkins-X <id> is owner/repository/branch, e.g. GET /vend/owner1/repo1/master -> 6
The backend logic for managing/generating build numbers is exactly as per `jx step next-buildnumber`, which is
using PipelineActivity CRDs.

I think it makes sense for this to eventually move _out_ of the jx command line tool, but that's where it is for now.