bug(install): Fix 'Brew' pre-req installation.
If 'brew' was not installed, then "jx install" attempts to install it, but fails because the command uses shell expansion,
but no shell. Fixed by changing the command to be 'sh', and have that run the brew installation command.
(I guess as pretty much everyone already has brew this doesn't get exercised too often!)