fix(applications): Added --org flag when removing app
This is needed because it's lost when importing an application.
I will submit a subsequent PR to store this info in a new application CRD
when a `create quickstart` or `import` is performed and use this to
get the org when deleting.

Note: At present, an application name is the same as the repo name, so
you can't have two different applications (repos) with the same name
but from different orgs. But TBH, that would be a daft setup anyway