bug(controller buildnumbers): Remove encoding of PipelineActivity.Name
Previously added some encoding of upper case letters in org/repo/branch to form legal K8S Name values, but this isn't working
properly. Removing that for now, so that the workaround in Prow (to ToLower everything) can be removed - as ultimately this is
causing occasional duplicate pipeline refs in PipelineActivities.