fix(ingress): hook exposeUrl was wiped if cert-manager failed
If the cert-manager deployment failed, the exposeUrl (and thus ingress)
for hook got wiped. This lead to `jx upgrade ingress` failing to update
github webooks because it thought the _old_ URL (it was trying to upgrade
from) was '/hook' rather than 'http://my.jx.url/hook'

By doing the `ensureCertmanagerSetup()` before modifying doing the delete/
recreate service annotations, we don't end up with an error in the
cert-manager installation causing data to get wiped but not re-created.