chore(users) users need to be identified by their login not email
this changes the behaviour introduced in #2669 which before that change
was incositent depending on if you where calling Create or
CreateOrUpdate.

Additionally FindByEmail would not have worked correctly previously due
to the fact that some users where based on their login and some based on
their email.  I found no useage of this code, so rather than fixing it I
have removed it.  If it transpires there are users of this in some
private repo then we can reinstate a working version.

see https://github.com/jenkins-x/jx/pull/2669#discussion_r24863822 for
the discussion