bug: update activities to complete on PR merge.
This is basically copying the workaround for issue #2410 from `PromoteOptions.waitForGitOpsPullRequest()` into
the workflow controller so that PipelineActivities get updated when the promotion PR merges. Normally, this is
done via `jx promote` during the build, but if that fails/times out then this ensures the workflow controller
tidies up the status to reflect that the PR is merged & promotion is complete.