feat(kpipelines): Add PipelineStructure CRD
This will allow us to record the syntax-based structure of a pipeline,
including nested and parallel stages, so that we can then map from a
running Pod to the stage (and possible parent stages) it belongs
to. This can then be used to populate PipelineActivity properly, and
to get logs correctly.

Note that this mapping can't actually be implemented until we move to
a current version of Build Pipeline in the chart we install - once
that's done, I'll follow this up with a change to actually perform
that mapping.

Also added helper methods for creating ParsedPipeline and
PipelineStructure in tests. Existing tests are only partially
converted currently - that'll be updated shortly.