fix multi context tekton status updates
Custom labels should only be applied to `PipelineRun` resources as these are unique per run and they contain information specific to that build.  If we add the labels to the Pipeline then we overwrite existing ones.  We were seeing the `prowJobName` label being updated on a single static `Pipeline` which in turn meant Tekton synced back that label onto every `PipelineRun`.

TL;DR this fixes the multi context status updates on GitHub