chore: Switch to sigs.k8s.io/yaml
This will let us incorporate k8s and Tekton types directly into
jenkins-x.yml - previously, we'd have trouble with that due to the use
of JSON struct tags in k8s/Tekton/etc, so that the default golang YAML
library would not unmarshal correctly.

This did require changing `loop_step/jenkins-x.yml` - a bare `- on` as
part of a string slice (or `foo: on` for a string field) would end up
unmarshalling as the string `true`, so yeah, `on`, `off`, `t`, `f`,
etc do need to be quoted if they're supposed to result in strings now.