feat: Allow specifying container configuration in pipeline YAML
Adds a new `containerOptions` to both `pipeline` and `stage`
`options`, which takes a `Container`. All steps underneath that
context then start with that `Container`'s configuration, with
`pipeline` configuration merged into `stage` configuration (and parent
`stage` merged into child `stage`).

Down the road, we may choose to add this for steps as well, but let's
see if what this adds is sufficient first.

Fixes #3312, #3366