fix(tekton) make pipelines and tasks unique per-build
Otherwise, updating a Pipeline while a build is in progress
causes the running Pipeline to adjust the rest of its execution
to match the new state, which means the build would execute
partly using the old version and partly using the new version,
which is a little confusing in most cases but could be a
serious correctness issue depending on the nature of the changes.

Uniqueness is handled by adding build numbers to Pipelines and
Tasks. Build numbers a tracked per-branch using annotations on
SourceRepository resources.