fix unit tests given the extra default git-merge step we now automatically add to support prow batches
some of these test changes I've made don't look right so will need to review, but right now we've been using these new steps fine in the prod cluster.  I'd prefer to get this in so we're not using a preview version on our prod cluster anymore.

Action on me to followup with the pipeline team.