If the name of the chart is an absolute path then do not try and version it.
If a user is asking to load a chart with an absolute path then we should
expect that they know what they are doing.  This is the case for
installing with gitops.

On linux this would have just caused a warning about failing to load a
version - but on Windows this will error as the resulting path from
filepath.Join would be a bogus path ( eg c:\foo\bar\C:\wibble\bang)
which would rather than causing FileExists to return false it would
return an error as the path is completely invalid.
eg GetFileAttributesEx C:\Users\jnord\.jx\jenkins-x-versions\charts\C:\workarea\source\github\teilo-org\jenkins-x-dev-environment\env.yml:
       - The filename, directory name, or volume label syntax is incorrect.