chore: Initial work on unification of build pack and jenkins-x.yml syntax
Here specifically, we're getting rid of `PipelineAgent` and
`PipelineStep`, so that you can use the `jenkins-x.yml` syntax for
agents and steps within a build pack or for overriding/adding to a
build pack in `jenkins-x.yml`. The existing syntax still works in
build pack context, but can't be used when defining the whole pipeline
- you'll get validation errors telling you what you should do
instead.

More is likely to come - but since this is functional as is with no
backwards compatibility concerns that I can find, hey, might as well
toss it up.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>