feat: add support for `jx step git fork-and-clone`
This command will fork and then clone locally a git repo.
It will fork if the git repo is in a different org (or user)
than the current git user.

Also adds support to `jx create pull request` to commit and push the
code in the current directory. Together these two commands allow you
to clone a git repo, run something (e.g. a `Makefile`) and then create
a pull request with the changes.   Also adds `hack/generate-clients.sh` that will use these commands to
clone and create PRs for the various clients for Jenkins X.

Signed-off-by: Pete Muir <pmuir@bleepbleep.org.uk>