feat(tekton): Define default pipelines, override steps in defaults
This adds a new field to `pipelineConfig.pipelines`: `default`. That
contains a `ParsedPipeline`, like
`pipelineConfig.pipelines.release.pipeline`, and is used as the
default for all pipelines (i.e., `release`, `pullRequest`, `feature`)
in the project/build pack if one isn't specified there already. It can
be overridden/added to via `pipelineConfig.pipelines.overrides`, which
also now has the `type` field, taking `replace`, `before`, and
`after`, defaulting to `replace` - if it's `replace`, the target step
is replaced with the new step or steps, while if it's `before` or
`after`, the new step or steps are put either before or after the
target step, as appropriate.

Example syntax in `jenkins-x.yml` or a build pack's `pipeline.yaml`
for defining a default pipeline:

```
pipelineConfig:
  pipelines:
    default:
      agent:
        image: go
      stages:
        - name: First Stage
          steps:
            - command: echo
              args: ['hello', 'there']
              name: hello
        - name: Second Stage
          steps:
            - command: echo
              args: ['goodbye']
              name: goodbye
```

Example override syntax (works with build packs with pipelines defined
as `default`, as `pipelineConfig.pipelines.release.pipeline`, or
standard lifecycles like `pipelineConfig.pipelines.release.build` etc):

```
pipelineConfig:
  pipelines:
    overrides:
      - pipeline: release
        stage: First Stage
        name: hello
        step:
          command: echo before we say hello...
        type: before
```

-------------

feat(tekton): Add --effective-pipeline to step create task

This allows you to print out the "effective" pipeline before CRD
generation but after the build pack, overrides, etc have been applied.

-------------

feat(tekton): Allow removing whole stages/lifecycles and pipelines,
and replacing/prepending/appending replacement steps for a whole stage.

If `name` is not given for an `override`, the whole stage (if `stages`
is specified) or whole pipeline (if just `pipelines` is specified) is
affected - either removed if no `step` or `steps` is given, or, for
stages, the exsting steps are replaced/prepended/appended depending on `type`.

-------------

feat(tekton): Allow overriding agent on stage or pipeline from default Tekton pipelines

Specify an `agent` to override the existing `agent` for either a
`stage` or a `pipeline`. Can be specified along with `step` or
`steps`, or on its own.

```
pipelineConfig:
  pipelines:
    overrides:
      - pipeline: release
        stage: Some Stage
        agent:
          image: some-image
```

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>