fix: remove the 'jx-' prefix from application names (#4050)
It seems that the applications are prefixed with 'jx-' regardless of the namesapce
where they are installed. This can create confusion when for instance users are trying to
promote an application. They are often use the application with the jx prefix which
leads to an error, because the chart is published without the prefix.

Signed-off-by: Cosmin Cojocar <cosmin.cojocar@gmx.ch>