fix(tekton): Use a working and useful versions directory (#4064)
We *were* using the typoed `cmmon_versions`, but once that typo was
fixed, we weren't using anything. Changing to just use the
now-existant `common_versions` wasn't particularly desirable, since it
messed with expected volumes etc, so instead let's use
`stable_versions`, which matches up with the pod templates we define
for the test.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>