feat: switch to a self uploading covered binary
If called when compiled into coverage mode, this binary will call itself, adding the `test.coverprofile` argument, and, after the subprocess exits, upload the coverage to codecov. It also fixes the output to remove the coverage lines generated by go. This allows us to use it transparently in tests, only setting the `CODECOV_TOKEN` into the environment.

Signed-off-by: Pete Muir <pmuir@bleepbleep.org.uk>