fix(tekton): Allow setting pipeline timeout and set default on PipelineRun to 240 hours (#4251)
* fix(tekton): Set a default PipelineRun timeout of 10 days

Starting in Tekton Pipelines v0.4.0, the default timeout for
`TaskRun`s of 10 minutes when no timeout is specified is actually
enforced. The only way to specify a timeout for a `TaskRun` is by
setting a timeout on the `PipelineRun`, so until
https://github.com/tektoncd/pipeline/issues/978 and/or
https://github.com/tektoncd/pipeline/issues/979 are addressed, we
should just set a ridiculously high timeout on all `PipelineRun`s.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>

* fix(tekton): Allow configuring Pipeline timeout

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>