fix(tekton): Metapipeline should pass env/labels to generated pipeline
We should be passing the custom labels/env vars/trigger/service
account/source directory that we'll get from Prow (see
https://github.com/jenkins-x/test-infra/blob/pipelinerunner/prow/cmd/pipeline/controller.go#L731-L739,
though it's not actually passing anything but labels currently) on to
the generated pipeline, not just use them on the metapipeline itself.

This also deals with propagating the `prowJobName` label, since that
comes from Prow itself.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>