chore(tekton): Generate PipelineActivity key from info rather than PR name
With metapipelines, the `PipelineRun` name is going to be different
than for the generated `PipelineRun`, but we want to track both of
them together for `PipelineActivity`, `get build logs`, etc. So let's
make the `Name` field of a `PipelineRunInfo` be the same for both
cases, by using the org/repo/branch/context/build number info from the
`PipelineRun`'s labels, rather than just using the `PipelineRun` name
directly. The metapipeline and generated pipeline will match in the
label values.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>