fix(tekton): Metapipeline should pass env/labels to generated pipeline (#4430)
* fix(tekton): Metapipeline should pass env/labels to generated pipeline

We should be passing the custom labels/env vars/trigger/service
account/source directory that we'll get from Prow (see
https://github.com/jenkins-x/test-infra/blob/pipelinerunner/prow/cmd/pipeline/controller.go#L731-L739,
though it's not actually passing anything but labels currently) on to
the generated pipeline, not just use them on the metapipeline itself.

This also deals with propagating the `prowJobName` label, since that
comes from Prow itself.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>

* fix(tekton): Prepend "metapipeline-" rather than appending it.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>

* chore(tekton): Generate PipelineActivity key from info rather than PR name

With metapipelines, the `PipelineRun` name is going to be different
than for the generated `PipelineRun`, but we want to track both of
them together for `PipelineActivity`, `get build logs`, etc. So let's
make the `Name` field of a `PipelineRunInfo` be the same for both
cases, by using the org/repo/branch/context/build number info from the
`PipelineRun`'s labels, rather than just using the `PipelineRun` name
directly. The metapipeline and generated pipeline will match in the
label values.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>