fix: add a path to enable semantic releases
For now, let’s just add a `--semantic-release` flag to `jx step create task` and the `pipelinerunner` so that we can optionally use semantic-release as a `jx step next-version` strategy.

Once meta-pipeline is in, and stabilises, we can perhaps do the `jx step next-version` in there, and use it to customise the semantic release.