fix: various fixes for the fork / rebase logic
Previously we were using the origin remote as the upstream, which works fine if the users fork is correct (i.e. has the branch which is to be updated) but fails if it doesn’t (e.g. a different user runs the command). By switching to the `pull/<ID>/head` remote on the upstream, we’re guaranteed to always be able to do fetch the branch.

We were also always trying to update a PR, regardless of owner, which won’t work.

Finally, we need to handle the case that rebases don’t ignore
empty commits. This seems to only happen in some cases. This patch
adds the ability to look out for those errors and run `git rebase —skip` if they happen.