feat(tekton): Add support for specifying volumes to add to stage pods
fixes #4665

This will allow pipeline authors to add existing volumes to the pods
created for their stages, which can then be mounted via
`containerOptions`. Note that this will result in errors for volume that
don't exist, and could cause problems, depending on the volume's
configuration, with parallel stages using the same volume.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>