fix: refactor ForkAndPullRepo
Previously this function branched pretty much at the
start which meant that it was very easy to end up
with fixes inconsistently applied. It is now a
single branch with the different paths done as
fork-joins which should be much easier to maintain.

I’ve also added unit tests for the common paths

Contributes to #4789