fix(tekton): Make sure PipelineRun has a running or terminated pod before logging.
If the first pod for a `PipelineRun` is pending, we end up thinking
it's started when we're finding valid names to look for, but it's not
actually running yet, so the later check for whether the run has
running pods fails and treats it as a completed and garbage-collected
run. So let's check for a running step first.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>