fix: Don't do release prepare for bdd and tekton contexts, copy pipeline
We're trying to run the `tekton` and `bdd` contexts on post-submit,
but they've been failing forever due to there not being a `release`
pipeline in either context. But if there _had_ been `release`
pipelines, it would have tried to do new release versions for those
contexts! So let's copy the `pullRequest` pipeline (until we're on a
new enough version on prod that we can switch to `default` pipelines
instead) and set `noReleasePrepare: true` to prevent the version
number changing too.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>