fix: don’t set push urls as remotes
Previously the code would use the push url as the remote url. This is fine when run in a pipeline but less than ideal when run locally (as in `jx boot`.

With this change we set the remote to the clone URL and use the push URL directly when we do pushes.