fix: Get logs from pod(s) if possible if BuildLogsURL starts with http
Right now, on prod, if you try to run `jx get build logs` on a
completed run, you're out of luck: it sees `BuildLogsURL` set, but
it's set to an `https` URL (or maybe an `http` URL, I didn't actually
check) and so just fails out. If we're not allowing streaming of logs
from non-bucket sources in the first place, let's at least just skip
`http*` cases and go back to the pods.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>