fix: Fix gh-pages for log storage in a couple ways
A few things were just completely borked in using gh-pages for log
storage for a while now. If the repo didn't already have a gh-pages
branch, we tried to reclone without deleting the existing attempted
clone, which was bad, and then when there was a gh-pages, we tried to
push `HEAD` without specifying that we actually want to push it to
`gh-pages`. This fixes both of those things.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>