fix: gc activities shouldn't try to delete running activities
Also, it shouldn't assume that the list of activities is sorted, just
to be safe, based on weirdness seen on the prod cluster where the
oldest two activities for a PR were kept. Additionally, we don't need
to explicitly delete `PipelineRun`s any more - those are now properly
owned by the relevant `PipelineActivity` and will get cleaned up
automatically when we delete the activity.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>