test: add an option in the makefile to disable the test caching only when is set
The test caching seems to be disable by default in all test without any option to
turn it on when test are executed locally. This option will allow to keep the
test caching on and disable it explicitly only when required.

This will speed up the test execution at least when run locally.

Signed-off-by: Cosmin Cojocar <cosmin.cojocar@gmx.ch>