fix: do not log an error when a ingress is not found for a service
This method is used by jx open/jx get urls to list the available URLs/ingress resource. It's
more appropriate to log a debug message when not ingress is found for a service instead of an error.
This will reduce the noise in case of services which are not exposed from the cluster.

Signed-off-by: Cosmin Cojocar <cosmin.cojocar@gmx.ch>