fix: Report error for nil pipeline for expected kind
Previously, we'd just end up with a nil pointer, and that's...never
helpful. This will instead report an error for when there is no
pipeline of a given kind in the effective config.

Also, I reworked how we check for expected errors in the test, and
made `override.MatchesPipeline` case-insensitive, since we've got
`pullrequest` in some places and `pullRequest` in others. Sigh.

fixes #5183

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>