fix: Revert making all created repos private by default
Reverts #5284. As seen in #5507, we weren't ever actually testing this
correctly (due to `--git-public false` actually resulting in
`--git-public` being true and `false` being treated as additional
arguments to `jx create cluster`). Fixing that showed that
promote-to-production fails when using the GitHub user as the env repo
owner as well, which is problematic in and of itself, but more
immediately, we can't get tests to pass without creating a second user
to own the private env repos or an org that allows private repos. This
is blocking anything from getting merged in `jenkins-x-versions`,
which means more and more things can be slipping into the periodic PRs
there that could be introducing more failure cases we can't detect, etc.

So...we should revert and revisit.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>