fix: Update existing PR before pushing in `jx step create pr`
I just did a batch of experiments and verified that the `synchronize`
GitHub webhook event does get sent twice if you push to an existing
PR, and then within a sufficiently rapid timeframe, edit the PR. I'm
not sure whether it's the title or the description being changed
that's the issue (or both), but regardless, updating the PR's
title/description _before_ we push results in only a single
`synchronize` event being sent. Woo!

fixes #5446

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>