fix(activity): Don't panic for batch builds with PRs without activities
If a PR goes into a batch build after its most recent
`PipelineActivity`(s) have been GC'd, it will currently cause a panic
due to the list of `PipelineActivity`s for the commit and branch from
the pull refs not having any elements in it. Let's bypass that - if
there are no such activities, we don't bother adding batch info for
that particular batch pull ref.

fixes #5812

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>