fix(logs): Check if activity is done or if run is done to get stage count for logs
I prefer checking the run, but there is a rare edge case where that
won't be enough, so let's check the activity's status as well.

Also realized that the `controller_build_test` for a running
`PipelineRun` wasn't 100% accurate, though it makes no functional
difference. I cleaned it up anyway.

fixes #5851

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>