fix(pipeline): Add GIT_* env vars to non-pod template images
If a step uses an image that isn't a pod template alias, like `go` or
`maven`, and instead uses an image name, with or without tag, such as
`kaniko/executor` or `gcr.io/jenkinsxio/builder-go:2.0.902-250`, we
haven't been adding all the environment variables that pod templates
automatically include, meaning those env vars aren't set unless
they're explicitly specified in the `jenkins-x.yml` or build pack. So
far as I can tell, the only ones we're missing right now are
`GIT_AUTHOR_NAME` and friends. So this PR adds them.

Figuring out exactly the right value to use is a challenge. The pod
templates all have `jenkins-x-bot` and `jenkins-x@googlegroups.com` as
the name/email, but it feels like it'd be better to use
`PipelineUsername` and `PipelineUserEmail` from team settings...if
they're set. Since `PipelineUserEmail` was just added a month and a
half ago, we can't assume it's going to be set in existing installs,
even after updating, so we'll use those values if they're not empty,
and fall back on the defaults if needed.

I also added constants for the default values, because why not.

fixes #5615

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>