fix(boot): Try to update local boot clone to config ref from versions
This will check if the boot clone isn't using the ref the desired tag
points to, after cloning the boot config if needed. If the clone is
using a different ref, it will try to stash any local changes, update
the clone to use the appropriate ref, and unstash those local
changes. If there are errors in the unstash due to conflicts, it'll
report the `git stash pop` error message and tell the user what they
need to do to fix it.

fixes #5929

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>