fix: Ignore tide context for last commit status on PRs
This should hopefully get `jx promote` and `jx delete app` to stop
getting confused by `tide` contexts, blocking merging due to the
`tide` context being, obviously, `pending`, and also first in the
list. So this will make sure that we check the latest non-`tide`
context with a state instead.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>