fix(logs): Log metapipeline when generated pipeline is not yet running
Don't just log the metapipeline if it failed or is still running and
there is no generated `PipelineRun` yet - also log it if the generated
`PipelineRun` exists but it doesn't yet have running pods (which can
be the case for a number of reasons).

Had to tweak one test case because it didn't have status on its `PipelineRun`.
fixes #6007

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>