fix: Don't use config.DefaultVersionsURL outside of version stream cloning logic
We shouldn't have been doing this in the first place, but definitely
not now that we've changed `config.DefaultVersionsURL` to be
overridable, and we've got a standard API for cloning the version
stream. So change places that were previously using
`config.DefaultVersionsURL` to use an empty string instead, which will
result in the team settings-provided version stream URL being used instead.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>