chore: Add builder-go-nodejs, switch BDD tests to use that image
This will enable us to use Cypress tests within BDD tests.

We also add support for updating the `jenkins-x*.yml`s in
`jenkins-x-versions` to use the latest `builder-go-nodejs` image,
since we'll be switching them to use that builder as well. Once this
merges and gets through builders, I'll update the image used in the
version stream update cronjob.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>