fix: do not force external dns on GKE cluster with custom domain
Resolves: https://github.com/jenkins-x/jx/issues/6128

Prior to this commit, `jx step ceate install values` forced external dns when detecting a GKE cluster with a custom domain.
This behavior could not be prevented by the user, and would lead to a failure if the service account did not have sufficient rights to configure the external dns.

Signed-off-by: Aurélien Lambert <aure@olli-ai.com>