fix: log an warning when no helm secrets are found in vault
This method seems to be called for every helm apply but most of the
helm charts do not have secrets in vault. This was relying on the fact
that read from vault returned an empty object instead of a error when
no secret is found in vault. After fixing that, this stared to return
an error for every helm apply.

Signed-off-by: Cosmin Cojocar <cosmin.cojocar@gmx.ch>