fix: prevent vault upgrade to make changes to cloud environment when running in pipeline
When running in pipeline (non `jx boo`) we are not makeing any modifications to
the could resources. The assumption is that they are in place.

The reason is that the pipeline user does not have sufficient previleges to do so
and providing these previleges will weaken security.

If a rebuild/upgrade of the cloud resources is needed `jx boot` must be executed
locally with a user having enough previleges.

fixes #5867