fix: Use debug image and /busybox/sh for /kaniko/executor
/bin/sh doesn't exist on Kaniko's `executor` or `warmer` images, which
is why we've been special-casing them all along. But that ends up
creating a number of other problems. There is a `debug-...` image for
every `executor` image, with `/busybox/sh`, so let's try using the
`debug-...` image with `/busybox/sh -c` as the command when we see
`/kaniko/executor` as the step command.

This should also allow environment variables to be used in kaniko step arguments.

fixes #6197
fixes #6058

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>