refactor: remove redundant conditional
Further up we have `user.Login` and what this does is assume this
function will always have a `user` object passed in. That user object or
whatever is returned from Github is assigned to `gitUser` variable.

Therefore the `gitUser` variable can never be `nil`, so this error will
never get printed or reached.

This commit removes this conditional because it's redundant and will
never get run.