fix: Upgrade Tekton to 0.8.0, remove Knative Build
Note that the original plan was to upgrade to 0.9.x, but that turns
out to be...messy. Tekton 0.9.x depends on k8s 1.13.x at compile time,
which results in a bunch of mock/fake changes being needed on our
side. Going to 0.8.0, with its dependency on k8s 1.12.x, did require
some changes, but not as many.

And more importantly than changes here, going to k8s 1.13.x at compile
time results in breaking changes to compilation for Knative Build and
Serving, Cert Manager, and Bankvault. While some of these can be
solved by upgrading, Knative Build ended at 0.7.0 (which depends on
k8s 1.12.x, and which we're upgrading our dependency to here, even
though we also remove all the Knative Build usage here at the same
time), and even if we remove our `go.mod` dependency on Knative Build,
we then end up pulling in 0.2.0 thanks to our Prow
dependency. So...things need to be reworked in some janky ways to
upgrade all the way to 0.9.2.

This also removes Knative Build - the addon is left around for now,
but any code actually using it is gone. go.mod still references it,
because otherwise we'd pick up Knative Build 0.2.0 via depending on
Prow, which would just mix things up more. I'm working on a Prow
change to remove Knative Build there which will completely eradicate
Knative Build.

fixes #6317, #4230

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>