fix: Stop lowercasing environmentGitOwner value
Prow is case-sensitive when it comes to matching events for an
org/repo name combo to a configured org/repo name combo in its
configuration. So if you're using an org with upper-case letters in
its name as things stand before this change (let's say `SomeOrg`), your Prow config will end
up having `someorg/environment-whatever-staging`, but the hooks sent
from GitHub will be for `SomeOrg/environment-whatever-staging`, and
Prow will not find any actions to take based on the `SomeOrg` event.

fixes #5914

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>