fix: Creating dev env repo should come from GitHub, even with different target providers
The default upstream boot config repo that's used as the basis for
creating the dev env repo for a new cluster is, at least as of now,
always on github.com, but `DuplicateGitRepoFromCommitish` only uses a
single `GitProvider`, meaning that if you're using GitHub Enterprise,
or Gitlab, etc, the attempt to get information needed about the
upstream repo in order to create the dev env repo tries to look for
the upstream repo's org/name in the provider that will have the dev
env repo in it. Which is bad.

So instead, let's check if the provider we'll be creating in is not
github.com and the upstream boot config repo we're going to use as the
basis for the dev env repo *is* on github.com, and if so, use a
different provider to get the required information.

fixes #6315

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>