fix: Fall back to anonymous GitHub provider for getting quickstarts
Similar problem as #6315 - if there are no GitHub credentials, we
still try to use an authenticated user for listing the quickstarts. So
let's watch for an error creating the git provider for listing
quickstarts from a location, and if that location is a default
location, use an anonymous GitHub provider instead.

fixes #6471

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>