fix: Fall back to not using provider for dev env creation in more cases
The specific issue prompting this was a GHE user with a custom boot
config upstream forked in github.com. That slipped through, since we
were specifically looking for cases where the destination wasn't
github.com and the origin was the default boot config repository, but
when I think about it more, it makes more sense to just fall back to
the simpler non-provider-driven approach to the origin for all cases
where the git host for the origin and destination repository URLs differ.

fixes #6551

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>