fix: Use requirements gitServer as default for import, create env
`opts.AddGitRepoOptionsArguments` defaults `--git-provider-url` to
`https://github.com`, no matter what the actual default should be for
the cluster. This makes sense for some of the cases where it's used
(i.e., `jx install`, `jx create cluster`, `jx create terraform`, etc),
but not for others (`jx import` (and therefore `jx create
quickstart`), `jx edit env`, and `jx create env`). In those cases, we
should default it to the `gitServer` in the team settings.

Except we don't actually set that with `jx boot`! It just stays on
`https://github.com`. So we also need to set that from the requirements.

fixes #6579

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>