fix: Lookup git service URL kind from jx-auth-config as well
We're looking for this in a simple set of patterns, secrets, and
`GitServices`, but once the `jx-auth-config` configmap is present
(i.e., via a combination of boot/vault/gitops), we won't be able to
find the kind for something that doesn't match one of the patterns we
check first, since the `jx-pipeline-git-` secrets won't exist any
more. So if we're in a cluster with `jx-auth-config`, let's look up
there too, before we check secrets or `GitServices`.

fixes #6589

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>