fix: allow `jx boot --no-update-git` to not mess with git
when testing out a possible PR on the git boot repository referenced in the `jx-requirements.yml` even using `--no-update-git` I was finding that boot would:

* lookup my local fork of the git repository and since its not in the version stream assume 'master' was the branch I wanted (despite being in a fork/branch already
* re-apply all changes from master to my local checkout causing merge hell
* then `jx step verify env` tries to push to master the local changes merged with upstream causing merge conflicts etc.

This PR fixes this by disabling the bit of `jx boot` which switches from the current gitRef -> `master` if the `--no-update-git` is provided. That flag also passes in an env var so that `jx step verify env` doesn't try to push local changes to master of the dev repo (which is only really intended for real installs from the real upstream git repo - not from local forks trying changes out in boot config).

There's an additional minor change for helmfile to disable lazily creating the dev Environment if helmfile is used; which breaks helmfile/helm 3 which wants to create/update that resource via helm

Signed-off-by: James Strachan <james.strachan@gmail.com>