fix: Get structure for rerun PipelineRun properly
If Lighthouse has rerun a `PipelineRun` due to our buddy the Tekton
race condition, the new `PipelineRun` has a different name than the
original. This means that there's no `PipelineStructure` with the same
name, so logging and the build controller don't know what to do with
it. So when we're looking up `PipelineStructure`s for a `PipelineRun`,
first try to look by the `PipelineRun` name, and then by the
`Pipeline` name associated with the `PipelineRun`. The `Pipeline` name
will always be the same for the rerun `PipelineRun` as it was for the
original, so this will work.

fixes #6646

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>