fix: Inherit build pack environment into pipelines
I missed this and feel silly - if you declare `env` variables at the
top level of a build pack, they won't get inherited properly either by
build packs that extend it or by pipelines that use it. So when we
extend build packs/pipelines, combine the `env` field as well as
agent, directory, and container options from the top level.

fixes #6668

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>