fix: Fetch k8s.io openapi dependencies as modules
https://github.com/googleapis/gnostic/commit/896953e6749863beec38e27029c804e88c3144b8
broke at least `k8s.io/apimachinery`'s dependency on
`github.com/googleapis/gnostic/OpenAPIv2`, so we need to fetch at
least `k8s.io/apimachinery` as modules to get it to pick up the right
version of `github.com/googleapis/gnostic`, and just to be safe, I did
the same for `k8s.io/api` as well.

I am not...happy with this, but it seems to work and doesn't seem to
have side effects that I can see.

fixes #6679

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>