fix: Replace dir placeholders on build pack stages
Missed this earlier in #6700. Also, when we're converting from a build
pack using `jenkins-x.yml` syntax, set the stage directory when
unset to the default if the stage contains steps, so we don't need to
put a redundant directory on each step in that stage. This logic could
probably be improved a bit, but it generates effective syntax that
works and is what I would expect it to be.

fixes #6717

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>