fix: Use kube crd name to look up the docker registry image
Prior to this commit, we'd use the kubernetes compliant crd name to
create the docker registry destination. But then later on in our code,
we'd use the Git repository name to look up the docker registry image
and it would fail.

It caused the following issue for some of our customers [1].

1.) https://github.com/jenkins-x/jx/issues/5322