fix: Don't pass target URL to GH status if not an http/https URL
Otherwise you get an error because the target URL has to be an
http/https URL. Sigh. So if the resulting target URL from the template
doesn't start with `http://` or `https://`, don't pass a target URL at all.

And yes, this all really needs to be redone. Still in the land of
tactical fixes here.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>